Updated main project layout
