+++ .gitignore
@@ -63,16 +63,6 @@ typings/
 
 .rpt2_cache
 
-/runtime
-
-/helpers/*.js
-/helpers/*.js.map
-/helpers/*.d.ts
-
-/compiler/*.js
-/compiler/*.js.map
-/compiler/*.d.ts
-
-/transitions/*.js
-/transitions/*.js.map
-/transitions/*.d.ts
+/*.js
+/*.js.map
+/*.d.ts
