--- .travis.yml
@@ -1,9 +1,0 @@
-language: node_js
-sudo: false
-node_js:
-  - "8"
-  - "10"
-install:
-  - npm install lerna --no-save
-script:
-  - npx lerna bootstrap
