--- compiler/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2017",
-    "sourceMap": true,
-    "noUnusedLocals": true,
-    "alwaysStrict": true,
-    "noImplicitThis": true,
-    "module": "commonjs",
-    "outDir": "./",
-    "declaration": true
-  },
-  "files": ["./index.ts"]
-}
