+++ packages/template-compiler/package.json
@@ -32,7 +32,7 @@
     "mocha": "^9.0.3",
     "rollup": "^2.55.1",
     "ts-node": "^10.1.0",
-    "typescript": "^4.3.5"
+    "typescript": "4.3.5"
   },
   "dependencies": {
     "@endorphinjs/template-parser": "^0.9.0",
