+++ src/index.ts
@@ -1,0 +1,7 @@
+import endorphin from '@endorphinjs/template-runtime';
+
+export default endorphin;
+export {
+    type Component, type ComponentDefinition, type Changes, type TweenOptions, type TweenFactory,
+    createComponent, mountComponent, unmountComponent, afterUpdate, composeTween, Store,
+} from '@endorphinjs/template-runtime';
