+++ src/runtime.ts
@@ -1,0 +1,2 @@
+export * from '@endorphinjs/template-runtime';
+export { default } from '@endorphinjs/template-runtime';
