--- transitions/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2017",
-    "sourceMap": true,
-    "noUnusedLocals": true,
-    "alwaysStrict": true,
-    "noImplicitThis": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "outDir": "./",
-    "declaration": true
-  },
-  "files": [
-    "./index.ts"
-  ]
-}
