+++ tsconfig.compiler.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "CommonJS"
+  },
+  "files": [
+    "./src/compiler.ts"
+  ]
+}
